requirements.txt
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
msgpack==1.1.0
PyJWT==2.10.1
redis==5.2.1
sqlparse==0.5.3
tzdata==2025.1
